+++ packages/site/src/components/getRuleForPlugin.ts
@@ -5,7 +5,8 @@ import { node } from "@flint.fyi/node";
 import { performance } from "@flint.fyi/performance";
 import { flint } from "@flint.fyi/plugin-flint";
 import { spelling } from "@flint.fyi/spelling";
-import { type AnyRule, json, md, ts, yaml } from "flint";
+import { vitest } from "@flint.fyi/vitest";
+import { type AnyRule, json, md, packageJson, ts, yaml } from "flint";
 
 const plugins = {
 	astro,
@@ -15,9 +16,11 @@ const plugins = {
 	jsx,
 	md,
 	node,
+	"package-json": packageJson,
 	performance,
 	spelling,
 	ts,
+	vitest,
 	yaml,
 };
 
@@ -33,7 +36,7 @@ export function getRuleForPlugin(pluginId: string, ruleId: string): AnyRule {
 		throw new Error(`Unknown rule for ${pluginId}: ${ruleId}.`);
 	}
 
-	return rule as AnyRule;
+	return rule;
 }
 
 export function getRuleForPluginSafe(
@@ -51,5 +54,5 @@ export function getRuleForPluginSafe(
 		return undefined;
 	}
 
-	return rule as AnyRule;
+	return rule;
 }